Original Files
I've attached the source code. It's currently working for just a single
computer. I need to take a look at the MPI stuff and our set up to see
how it needs to change so it can write to a shared hard drive. I've
tried to comment it as best as I could. If you have questions, please
let me know.

I also still need to work on converting the strings into integers, but
that shouldn't take too much longer. And it's not set up just yet to
reallocate memory if there are more samples / alleles than the space
provided.

For reading the file, the file must be in the format specified by
"samples.txt".
(This can be changed if we want to format it differently.)

The function will "return" (read the code to see what I mean by return):

how many samples there are (as an integers),
the "names" of each sample (as a 1D array of strings),
how many loci there are (as a 1D array of integers),
the names of each loci (as a 2D array of strings),
how many possible allele pairings there are (as a 2D array of integers),
and then all the possible allele pairings(as a 3D array of integers).

Because of this, I imagine it would be easiest to have the master
process send 1 single sample to each slave at a time, but I can change
my code as needed.

For writing the actual pairing possibilities, it's currently set up to
take a single sample, and write all of the possibilities. See the source
code for details.

NOTE: For the possible pairing arrays, the code is currently set up such
that the pairs are stored as 2 separate, but consecutive elements.

~Julian