Parallel Processing of Allele Combinations
Justin Jiang: "I'm sending you my code that I have so far. I think the
general structure is correct but I was having some issues with the data
types. If I send arrays of integers to the slaves, I am not sure what
they should be sending back to the master. If it is a string, then I am
not sure how to convert from int to string and send it back so if anyone
could help with that I would appreciate it. Once that is cleared up I
can finish with the printing/details and do the makefile and whatever
else." (11/08/13)