Current version as of 11/11/13
Justin Jiang:
"here are the current versions of the MPI program and the Python
progress bar. Also, I think something happened with my flash drive when
I was copying the Python file onto it so that may be the reason why the
progress bar wouldn't work on the cluster when we were testing today.
This version should be working properly so whoever is in school tomorrow
should probably try testing it again with this file." (11/11/13)